refactor: move OlapDictionary data model to framework/core/infrastructure/dictionary.rs
Following the Table pattern: the infra-agnostic data model (OlapDictionary,
DictionaryColumn, DictionaryLayout, DictionaryLifetime, DictionarySource,
all external source types) now lives in framework/core/infrastructure/dictionary.rs.

infrastructure/olap/clickhouse/dictionary.rs becomes an 8-line re-export shim
so all existing callers at the clickhouse path continue to work unchanged.

Addresses Callicles' review: "this is very explicitly Clickhouse-specific [so
it belongs here]" — the data model no longer lives alongside the DDL execution
code.
