+++ apps/framework-cli/src/framework/core/infra_delta.rs
@@ -1,3 +1,4 @@
+use crate::framework::core::infrastructure::dictionary::OlapDictionary;
 use crate::framework::core::infrastructure::materialized_view::MaterializedView;
 use crate::framework::core::infrastructure::select_row_policy::SelectRowPolicy;
 use crate::framework::core::infrastructure::sql_resource::SqlResource;
@@ -8,7 +9,6 @@ use crate::framework::core::infrastructure::view::{Dmv1View, View};
 use crate::framework::core::infrastructure_map::{
     Change, ColumnChange, InfrastructureMap, OlapChange, TableChange,
 };
-use crate::infrastructure::olap::clickhouse::dictionary::OlapDictionary;
 use crate::infrastructure::olap::ddl_ordering::{AtomicOlapOperation, DependencyInfo};
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
